os: Start tty services in the end
Move 'console-font-service' and 'mingetty-service' instances to the
beginning of the 'services' list to make them start after the other
services.

Specifically this is needed because of auto-login to tty1: my
".bash_profile" starts dmd which runs irexec (among other daemons) which
relies on a running lircd, so 'lirc-service' should be started before
running auto-login.