Remove scripts
Move "profile.scm" and "system.scm" to my "guile-config" repository.

Remove "usr-bin.sh" as it is useless (as pointed in "help" of my
"config" repository I can just use a one-liner for making a proper
symlink).

Remove "guix-builders.sh" as it is just an excerpt from the
manual (maybe it will appear in my future "shell-scripts" repository).