os: Replace deprecated 'console-keymap-service'
It was deprecated by
<http://git.savannah.gnu.org/cgit/guix.git/commit/?id=3a665637afc32a142dc24a77ce7ce9235eb6a3af>,
and apparently, it is going to be removed.

It is always nice to see when a feature added by you (see
<http://git.savannah.gnu.org/cgit/guix.git/commit/?id=5eca94594d5f0d834d4ca918b894400e3a7f6aa1>
and
<http://git.savannah.gnu.org/cgit/guix.git/commit/?id=b3d05f487014b1c34c45b8c693a974500e689603>)
is replaced by something inappropriate: it was replaced by
'keyboard-layout' which tries to set a layout everywhere (in console, in
X and even in grub).  And this is exactly what I do not need.

See also discussion around
<https://lists.gnu.org/archive/html/guix-devel/2019-04/msg00176.html>.

* modules/al/guix/services/linux.scm (loadkeys-service-type): New service type.
* system-config/os-main.scm: Use it instead of 'console-keymap-service'.
* system-config/os-32-to-64.scm: Likewise.
* system-config/bare-bones.scm: Likewise.